Added Fabric Controller with views using EF scaffolding

+++ MyFabricTrackerWebApp/Models/MainCategory.cs
@@ -7,7 +7,7 @@ namespace MyFabricTrackerWebApp.Models
 {
 	public class MainCategory
 	{
-		public int MainCategoryId { get; set; }
+		public long MainCategoryId { get; set; }
 		public string MainCategoryName { get; set; }
 		public List<SubCategory> SubCategories { get; set; }
