+++ MyFabricTrackerWebApp/Models/Transaction.cs
@@ -7,8 +7,8 @@ namespace MyFabricTrackerWebApp.Models
 {
 	public class Transaction
 	{
-		public int TransactionId { get; set; }
-		public int FabricId { get; set; }
+		public long TransactionId { get; set; }
+		public long FabricId { get; set; }
 		public virtual Fabric Fabric { get; set; }
 		public int? InchesQty { get; set; }
 		public int? FatQtrQty { get; set; }
